lwiplinux-4.11

+++ drivers/gpu/drm/omapdrm/omap_dmm_tiler.c
@@ -388,7 +388,7 @@ struct tiler_block *tiler_reserve_2d(enum tiler_fmt fmt, uint16_t w,
 	u32 min_align = 128;
 	int ret;
 	unsigned long flags;
-	u32 slot_bytes;
+	size_t slot_bytes;
 
 	BUG_ON(!validfmt(fmt));
